Dropping DB driver system for com_finder fixing several issues, esp Postgres
